requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
Django-Verify-Email==2.0.3
idna==3.4
kombu==5.3.3
Pillow==10.1.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
